requirements.txt
# local package
-e .
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
jupyter

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
